Update the documentation
