+++ matfree/lanczos.py
@@ -1,27 +1,12 @@
-"""All things Lanczos' algorithm.
+"""Lanczos-style matrix decompositions.
 
-This includes
-stochastic Lanczos quadrature (extending the integrands
-in [hutchinson][matfree.hutchinson] to those that implement
-stochastic Lanczos quadrature),
-Lanczos-implementations of matrix-function-vector products,
-and various Lanczos-decompositions of matrices.
+This module includes various Lanczos-decompositions of matrices
+(tridiagonalisation, bidiagonalisation, etc.).
 
-Examples
---------
->>> import jax.random
->>> import jax.numpy as jnp
->>>
->>> jnp.set_printoptions(1)
->>>
->>> M = jax.random.normal(jax.random.PRNGKey(1), shape=(10, 10))
->>> A = M.T @ M
->>> v = jax.random.normal(jax.random.PRNGKey(2), shape=(10,))
->>>
->>> # Compute a matrix-logarithm with Lanczos' algorithm
->>> matfun_vec = funm_vector_product_spd(jnp.log, 4, lambda s: A @ s)
->>> matfun_vec(v)
-Array([-4. , -2.1, -2.7, -1.9, -1.3, -3.5, -0.5, -0.1,  0.3,  1.5],      dtype=float32)
+For stochastic Lanczos quadrature, see
+[matfree.funm_trace][matfree.funm_trace].
+For matrix-function-vector products, see
+[matfree.funm][matfree.funm].
 """
 
 from matfree.backend import containers, control_flow, func, linalg, np
